+++ src/main/java/dream/fcard/MainApp.java
@@ -17,8 +17,8 @@ import dream.fcard.model.ModelManager;
 import dream.fcard.model.ReadOnlyAddressBook;
 import dream.fcard.model.ReadOnlyUserPrefs;
 import dream.fcard.model.UserPrefs;
-import dream.fcard.ui.Ui;
-import dream.fcard.ui.UiManager;
+import dream.fcard.gui.Ui;
+import dream.fcard.gui.UiManager;
 import javafx.application.Application;
 import javafx.stage.Stage;
 import dream.fcard.commons.exceptions.DataConversionException;
