+++ src/main/java/dream/fcard/gui/CommandBox.java
@@ -1,4 +1,4 @@
-package dream.fcard.ui;
+package dream.fcard.gui;
 
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
