+++ src/main/java/dream/fcard/gui/StatusBarFooter.java
@@ -1,4 +1,4 @@
-package dream.fcard.ui;
+package dream.fcard.gui;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
