+++ src/main/java/dream/fcard/gui/Ui.java
@@ -1,4 +1,4 @@
-package dream.fcard.ui;
+package dream.fcard.gui;
 
 import javafx.stage.Stage;
