+++ src/main/java/dream/fcard/model/cards/Person.java
@@ -1,4 +1,4 @@
-package dream.fcard.model.person;
+package dream.fcard.model.cards;
 
 import static dream.fcard.commons.util.CollectionUtil.requireAllNonNull;
