+++ src/main/java/dream/fcard/model/cards/exceptions/DuplicatePersonException.java
@@ -1,4 +1,4 @@
-package dream.fcard.model.person.exceptions;
+package dream.fcard.model.cards.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Persons (Persons are considered duplicates if they have the same
